requirements.txt
flask
Jinja2
pytest
pyyaml
google-cloud-aiplatform
firebase-admin
langchain
langchain_community